+++ src/jsMain/kotlin/com/wadii/domain/model/call/CallChannel.kt
@@ -1,6 +1,6 @@
 package com.wadii.domain.model.call
 
 object CallChannel {
-    fun name(userId: Int, contactId: Int): String =
+    fun name(userId: Long, contactId: Long): String =
         listOf(userId, contactId).sorted().joinToString("_")
 }
